refactor: add mb prop to CategoryNav component and move types to shared folder
