+++ client/src/components/CategoryNav/index.ts
@@ -1,0 +1,3 @@
+import CategoryNav from './CategoryNav';
+
+export default CategoryNav;
